remove firebase
