+++ .env.example
@@ -1,8 +1,3 @@
 NODE_ENV=XXXXX
 SITE_NAME=XXXXX
-API_KEY=XXXXX
-PROJECT_ID=XXXXX
-SENDER_ID=XXXXX
-APP_ID=XXXXX
-MEASUREMENT_ID=XXXXX
 API_PATH=XXXXX
